requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==3.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.7
django-annoying==0.10.6
django-appconf==1.0.4
django-cors-headers==3.4.0
django-dbbackup==3.3.0
django-imagekit==4.0.2
django-import-export==2.3.0
django-mysql==3.8.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.0.1
flake8==3.8.3
idna==2.10
importlib-metadata==1.7.0
jdcal==1.4.1
MarkupPy==1.14
mccabe==0.6.1
mysqlclient==2.0.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
pathspec==0.8.0
pilkit==2.0
Pillow==7.2.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rest-social-auth==4.2.0
six==1.15.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
